+++ packages/app-solana/src/app.ts
@@ -1,0 +1,41 @@
+import { IDeviceConnection } from '@cypherock/sdk-interfaces';
+import { SDK } from '@cypherock/sdk-core';
+
+import * as operations from './operations';
+
+export class SolanaApp {
+  private readonly sdk: SDK;
+
+  private static readonly APPLET_ID = 10;
+
+  private constructor(sdk: SDK) {
+    this.sdk = sdk;
+  }
+
+  public static async create(connection: IDeviceConnection) {
+    const sdk = await SDK.create(connection, SolanaApp.APPLET_ID);
+    return new SolanaApp(sdk);
+  }
+
+  public async getPublicKeys(params: operations.IGetPublicKeysParams) {
+    return this.sdk.runOperation(() =>
+      operations.getPublicKeys(this.sdk, params),
+    );
+  }
+
+  public async getUserVerifiedPublicKey(
+    params: operations.IGetUserVerifiedPublicKeyParams,
+  ) {
+    return this.sdk.runOperation(() =>
+      operations.getUserVerifiedPublicKey(this.sdk, params),
+    );
+  }
+
+  public async destroy() {
+    return this.sdk.destroy();
+  }
+
+  public async abort() {
+    await this.sdk.sendAbort();
+  }
+}