+++ apps/node/CHANGELOG.md
@@ -1,5 +1,14 @@
 # nodejs-test
 
+## 0.0.40
+
+### Patch Changes
+
+- [#90](https://github.com/Cypherock/sdk/pull/90) [`6aa86ea`](https://github.com/Cypherock/sdk/commit/6aa86ea7f643a503a82bd7fe6e2bc00b89dc9ec6) Thanks [@irshadCypherock](https://github.com/irshadCypherock)! - added solana getPublicKeys/getUserVerifiedPublicKey
+
+- Updated dependencies [[`6aa86ea`](https://github.com/Cypherock/sdk/commit/6aa86ea7f643a503a82bd7fe6e2bc00b89dc9ec6)]:
+  - @cypherock/sdk-app-solana@0.0.1
+
 ## 0.0.39
 
 ### Patch Changes